Access priority sorter plugin info via short name
In [version 2.0](https://github.com/jenkinsci/priority-sorter-plugin/commit/8f2ae302bd98a3e7edd66f75bed6324ab85c57a4)
of the priority sorter plugin, the long name of the plugin was renamed
to "Jenkins Priority Sorter Plugin", which broke the `priority_sorter`
function in the `properties` module.

The short name did not change, so we now use that instead to ensure
that we generate a configuration that matches the latest required by
the plugin.

Change-Id: I89a8fdabeac63ea482b3b1bd8ddf2dd83c1c7435
Task: 3112
Story: 2000655